refactor(settings): 设置页收敛，外观与关于收进二级页
- 新增「外观」二级页：深色模式 + 主题色迁入，一级页收为单行入口（色点展示当前主题色）
- 新增「关于我们」二级页：官网、联系作者、捐助我们（含对话框）、隐私政策迁入
- 「摘要语言」独立条目并入「语言设置」行：行点击弹说明，trailing 仍为语言切换
- 「用系统浏览器打开外链」去掉说明对话框，关键信息并入 supporting 文案，行点击即切换开关
- 「版本」与「检查更新」合并为一行：点击检查（apk 渠道）/ 跳官网（iOS），play 渠道仅展示版本号
- 清理不再使用的词条：check_updates、language_system_follow、open_links_in_browser_message
