+++ shared/src/commonMain/composeResources/values-zh/strings.xml
@@ -9,9 +9,9 @@
     <string name="theme_follow_system">跟随系统</string>
     <string name="theme_light">浅色</string>
     <string name="theme_dark">深色</string>
+    <string name="appearance">外观</string>
     <string name="app_settings">应用设置</string>
     <string name="language_settings">语言设置</string>
-    <string name="language_system_follow">跟随系统语言</string>
     <string name="language_option_follow_system">跟随系统</string>
     <string name="language_option_chinese">中文</string>
     <string name="language_option_english">英文</string>
@@ -26,17 +26,17 @@
     <string name="summary_lang_capture_label">语言</string>
     <string name="summary_lang_capture_identity_note">提交时会带上你的 GitHub 身份，方便我们后续通知你。</string>
     <string name="open_links_in_browser">用系统浏览器打开外链</string>
-    <string name="open_links_in_browser_desc">关闭后优先使用内置浏览器 · 点击了解</string>
-    <string name="open_links_in_browser_message">外部链接交给系统浏览器打开，可正常通过 Product Hunt 等网站的人机验证，并支持翻译、自动填充和已有登录态。\n\n关闭后，外链优先使用内置浏览器打开。\n\nGitHub README 阅读不经外链流程，不受此设置影响。</string>
+    <string name="open_links_in_browser_desc">系统浏览器可通过人机验证、复用登录态 · 关闭后优先使用内置浏览器</string>
     <string name="close">关闭</string>
     <string name="datasource_settings">数据源设置</string>
     <string name="current_datasource">当前使用: Gemini API</string>
     <string name="about">关于</string>
-    <string name="check_updates">版本更新</string>
     <string name="version_up_to_date">已是最新</string>
     <string name="version">版本</string>
+    <string name="version_tap_to_check">点击检查更新</string>
     <string name="about_us">关于我们</string>
     <string name="about_us_desc">了解更多关于 Trending AI 的信息</string>
+    <string name="official_website">官网</string>
     <string name="contact_author">联系作者</string>
     <string name="donate">捐助我们</string>
     <string name="donate_message">我们利用有限的业余时间打造 Trending AI，希望帮你更高效地发现全球技术热点。\n如果它对你有帮助，欢迎捐赠支持我们。\n\n所有捐赠都将用于改善服务器与运营，让产品持续前行。</string>
