+++ shared/src/commonMain/composeResources/values/strings.xml
@@ -9,9 +9,9 @@
     <string name="theme_follow_system">System</string>
     <string name="theme_light">Light</string>
     <string name="theme_dark">Dark</string>
+    <string name="appearance">Appearance</string>
     <string name="app_settings">App Settings</string>
     <string name="language_settings">Language</string>
-    <string name="language_system_follow">Follow system language</string>
     <string name="language_option_follow_system">Follow system</string>
     <string name="language_option_chinese">Chinese</string>
     <string name="language_option_english">English</string>
@@ -26,17 +26,17 @@
     <string name="summary_lang_capture_label">Language</string>
     <string name="summary_lang_capture_identity_note">We&apos;ll include your GitHub identity so we can follow up with you.</string>
     <string name="open_links_in_browser">Open links in browser</string>
-    <string name="open_links_in_browser_desc">Turn off to prefer the built-in browser · Tap to learn more</string>
-    <string name="open_links_in_browser_message">External links open in the system browser, which passes human verification on sites like Product Hunt and supports translation, autofill and existing logins.\n\nWhen turned off, external links prefer the built-in browser.\n\nGitHub README reading does not go through external link handling and is not affected by this setting.</string>
+    <string name="open_links_in_browser_desc">System browser passes human verification and keeps your logins · Turn off to prefer the built-in browser</string>
     <string name="close">Close</string>
     <string name="datasource_settings">Data Source</string>
     <string name="current_datasource">Current: Gemini API</string>
     <string name="about">About</string>
-    <string name="check_updates">Version Update</string>
     <string name="version_up_to_date">Up to date</string>
     <string name="version">Version</string>
+    <string name="version_tap_to_check">Tap to check for updates</string>
     <string name="about_us">About Us</string>
     <string name="about_us_desc">Learn more about Trending AI</string>
+    <string name="official_website">Official Website</string>
     <string name="contact_author">Contact Author</string>
     <string name="donate">Donate</string>
     <string name="donate_message">We build Trending AI in our limited spare time, hoping to help you discover global tech trends more efficiently.\nIf you find it helpful, a donation would mean a lot to us.\n\nAll donations go toward improving our servers and operations to keep the product moving forward.</string>
