+++ shared/src/commonMain/kotlin/whl/trending/ai/ui/settings/AppearanceScreen.kt
@@ -1,0 +1,215 @@
+package whl.trending.ai.ui.settings
+
+import whl.trending.ai.data.local.DEFAULT_SEED_ARGB
+import whl.trending.ai.data.local.ThemeMode
+import whl.trending.ai.data.local.globalSettingsManager
+import whl.trending.ai.core.platform.trackEvent
+import whl.trending.ai.ui.theme.PRESET_PALETTE
+import whl.trending.ai.ui.theme.ThemeSeed
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.ExperimentalLayoutApi
+import androidx.compose.foundation.layout.FlowRow
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.selection.selectableGroup
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.verticalScroll
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.ArrowBack
+import androidx.compose.material.icons.filled.Check
+import androidx.compose.material.icons.filled.ColorLens
+import androidx.compose.material.icons.filled.Palette
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.SegmentedButton
+import androidx.compose.material3.SegmentedButtonDefaults
+import androidx.compose.material3.SingleChoiceSegmentedButtonRow
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.luminance
+import androidx.compose.ui.semantics.Role
+import androidx.compose.ui.semantics.contentDescription
+import androidx.compose.ui.semantics.role
+import androidx.compose.ui.semantics.semantics
+import androidx.compose.ui.unit.dp
+import org.jetbrains.compose.resources.stringResource
+import trendingai.shared.generated.resources.Res
+import trendingai.shared.generated.resources.appearance
+import trendingai.shared.generated.resources.back
+import trendingai.shared.generated.resources.dark_mode
+import trendingai.shared.generated.resources.theme_color
+import trendingai.shared.generated.resources.theme_dark
+import trendingai.shared.generated.resources.theme_follow_system
+import trendingai.shared.generated.resources.theme_light
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun AppearanceScreen(onBack: () -> Unit) {
+    val themeMode by globalSettingsManager.themeMode.collectAsState(ThemeMode.FOLLOW_SYSTEM)
+    val seedColor by globalSettingsManager.seedColor.collectAsState(DEFAULT_SEED_ARGB)
+
+    Scaffold(
+        topBar = {
+            TopAppBar(
+                title = { Text(stringResource(Res.string.appearance)) },
+                navigationIcon = {
+                    IconButton(onClick = onBack) {
+                        Icon(Icons.AutoMirrored.Filled.ArrowBack, contentDescription = stringResource(Res.string.back))
+                    }
+                }
+            )
+        }
+    ) { innerPadding ->
+        Column(
+            modifier = Modifier
+                .padding(innerPadding)
+                .fillMaxSize()
+                .verticalScroll(rememberScrollState())
+                .padding(horizontal = 16.dp, vertical = 12.dp)
+        ) {
+            Row(
+                verticalAlignment = Alignment.CenterVertically,
+                modifier = Modifier.padding(bottom = 12.dp)
+            ) {
+                Icon(
+                    Icons.Default.Palette,
+                    null,
+                    tint = MaterialTheme.colorScheme.onSurfaceVariant
+                )
+                Text(
+                    text = stringResource(Res.string.dark_mode),
+                    style = MaterialTheme.typography.bodyLarge,
+                    modifier = Modifier.padding(start = 16.dp)
+                )
+            }
+            SingleChoiceSegmentedButtonRow(
+                modifier = Modifier.fillMaxWidth()
+            ) {
+                ThemeMode.entries.forEachIndexed { index, mode ->
+                    val labelRes = when (mode) {
+                        ThemeMode.FOLLOW_SYSTEM -> Res.string.theme_follow_system
+                        ThemeMode.LIGHT -> Res.string.theme_light
+                        ThemeMode.DARK -> Res.string.theme_dark
+                    }
+                    SegmentedButton(
+                        selected = themeMode == mode,
+                        onClick = {
+                            trackEvent("settings_theme_change", mapOf("theme" to mode.name.lowercase()))
+                            globalSettingsManager.setThemeMode(mode)
+                        },
+                        shape = SegmentedButtonDefaults.itemShape(
+                            index = index,
+                            count = ThemeMode.entries.size
+                        ),
+                        label = {
+                            Text(
+                                text = stringResource(labelRes),
+                                style = MaterialTheme.typography.labelMedium
+                            )
+                        }
+                    )
+                }
+            }
+
+            Spacer(Modifier.height(20.dp))
+
+            Row(
+                verticalAlignment = Alignment.CenterVertically,
+                modifier = Modifier.padding(bottom = 12.dp)
+            ) {
+                Icon(
+                    Icons.Default.ColorLens,
+                    null,
+                    tint = MaterialTheme.colorScheme.onSurfaceVariant
+                )
+                Text(
+                    text = stringResource(Res.string.theme_color),
+                    style = MaterialTheme.typography.bodyLarge,
+                    modifier = Modifier.padding(start = 16.dp)
+                )
+            }
+            SwatchGrid(
+                selected = seedColor,
+                onSelect = { seed ->
+                    trackEvent("settings_seed_color", mapOf("seed" to seed.id))
+                    globalSettingsManager.setSeedColor(seed.argb)
+                }
+            )
+        }
+    }
+}
+
+@OptIn(ExperimentalLayoutApi::class)
+@Composable
+private fun SwatchGrid(
+    selected: Long,
+    onSelect: (ThemeSeed) -> Unit,
+) {
+    FlowRow(
+        modifier = Modifier
+            .fillMaxWidth()
+            .selectableGroup(),
+        horizontalArrangement = Arrangement.spacedBy(12.dp),
+        verticalArrangement = Arrangement.spacedBy(12.dp),
+    ) {
+        PRESET_PALETTE.forEach { seed ->
+            ThemeSwatch(
+                seed = seed,
+                selected = seed.argb == selected,
+                onClick = { onSelect(seed) },
+            )
+        }
+    }
+}
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+private fun ThemeSwatch(
+    seed: ThemeSeed,
+    selected: Boolean,
+    onClick: () -> Unit,
+) {
+    val color = Color(seed.argb)
+    val name = stringResource(seed.nameRes)
+    Surface(
+        selected = selected,
+        onClick = onClick,
+        shape = CircleShape,
+        color = color,
+        modifier = Modifier
+            .size(40.dp)
+            .semantics {
+                contentDescription = name
+                role = Role.RadioButton
+            },
+    ) {
+        if (selected) {
+            Box(contentAlignment = Alignment.Center, modifier = Modifier.fillMaxSize()) {
+                Icon(
+                    imageVector = Icons.Default.Check,
+                    contentDescription = null,
+                    tint = if (color.luminance() < 0.5f) Color.White else Color.Black,
+                )
+            }
+        }
+    }
+}
