+++ shared/src/commonMain/kotlin/whl/trending/ai/ui/settings/SettingsScreen.kt
@@ -17,20 +17,14 @@ import whl.trending.ai.auth.AuthState
 import whl.trending.ai.auth.globalAuthManager
 import whl.trending.ai.data.remote.ApiException
 import whl.trending.ai.data.repository.TrendingRepository
-import whl.trending.ai.ui.home.githubLogoPainter
 import whl.trending.ai.ui.home.HomeTab
 import whl.trending.ai.notification.globalDailyPicksNotifier
-import whl.trending.ai.ui.theme.PRESET_PALETTE
-import whl.trending.ai.ui.theme.ThemeSeed
 import whl.trending.ai.update.globalUpdateChecker
 
+import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.text.selection.SelectionContainer
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.ExperimentalLayoutApi
-import androidx.compose.foundation.layout.FlowRow
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
@@ -39,31 +33,23 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.selection.selectableGroup
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material3.ExperimentalMaterial3ExpressiveApi
 import androidx.compose.material3.LoadingIndicator
 import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
-import androidx.compose.material.icons.filled.AlternateEmail
-import androidx.compose.material.icons.filled.Check
-import androidx.compose.material.icons.filled.ColorLens
 import androidx.compose.material.icons.filled.Email
 import androidx.compose.material.icons.filled.Favorite
 import androidx.compose.material.icons.filled.FiberNew
 import androidx.compose.material.icons.filled.Home
 import androidx.compose.material.icons.filled.Info
-import androidx.compose.material.icons.filled.PrivacyTip
 import androidx.compose.material.icons.filled.Language
 import androidx.compose.material.icons.filled.Notifications
 import androidx.compose.material.icons.filled.Numbers
 import androidx.compose.material.icons.filled.OpenInBrowser
 import androidx.compose.material.icons.filled.Palette
-import androidx.compose.material.icons.filled.Refresh
-import androidx.compose.material.icons.filled.Translate
 import androidx.compose.material.icons.filled.Feedback
-import androidx.compose.material.icons.filled.VolunteerActivism
 import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.DropdownMenu
 import androidx.compose.material3.DropdownMenuItem
@@ -77,10 +63,6 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.material3.SnackbarHost
 import androidx.compose.material3.SnackbarHostState
 import androidx.compose.material3.SnackbarResult
-import androidx.compose.material3.SegmentedButton
-import androidx.compose.material3.SegmentedButtonDefaults
-import androidx.compose.material3.SingleChoiceSegmentedButtonRow
-import androidx.compose.material3.Surface
 import androidx.compose.material3.Switch
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextButton
@@ -96,33 +78,21 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.luminance
-import androidx.compose.ui.semantics.Role
-import androidx.compose.ui.semantics.contentDescription
-import androidx.compose.ui.semantics.role
-import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.unit.dp
-import org.jetbrains.compose.resources.painterResource
 import org.jetbrains.compose.resources.stringResource
 import trendingai.shared.generated.resources.Res
 import trendingai.shared.generated.resources.about
 import trendingai.shared.generated.resources.about_us
 import trendingai.shared.generated.resources.about_us_desc
-import trendingai.shared.generated.resources.confirm
-import trendingai.shared.generated.resources.contact_author
-import trendingai.shared.generated.resources.donate
-import trendingai.shared.generated.resources.donate_alipay
-import trendingai.shared.generated.resources.donate_github_desc
-import trendingai.shared.generated.resources.donate_message
+import trendingai.shared.generated.resources.appearance
 import trendingai.shared.generated.resources.app_settings
 import trendingai.shared.generated.resources.back
-import trendingai.shared.generated.resources.check_updates
 import trendingai.shared.generated.resources.close
 import trendingai.shared.generated.resources.daily_picks_notification
 import trendingai.shared.generated.resources.daily_picks_notification_desc
 import trendingai.shared.generated.resources.notification_permission_denied
-import trendingai.shared.generated.resources.dark_mode
 import trendingai.shared.generated.resources.default_home_tab
 import trendingai.shared.generated.resources.default_home_tab_desc
 import trendingai.shared.generated.resources.hackernews_title
@@ -132,16 +102,13 @@ import trendingai.shared.generated.resources.language_settings
 import trendingai.shared.generated.resources.language_option_chinese
 import trendingai.shared.generated.resources.language_option_english
 import trendingai.shared.generated.resources.language_option_follow_system
-import trendingai.shared.generated.resources.language_system_follow
 import trendingai.shared.generated.resources.new_only_default
 import trendingai.shared.generated.resources.new_only_default_desc
 import trendingai.shared.generated.resources.open_links_in_browser
 import trendingai.shared.generated.resources.open_links_in_browser_desc
-import trendingai.shared.generated.resources.open_links_in_browser_message
 import trendingai.shared.generated.resources.open_system_settings
 import trendingai.shared.generated.resources.personalization
 import trendingai.shared.generated.resources.settings
-import trendingai.shared.generated.resources.theme_color
 import trendingai.shared.generated.resources.theme_dark
 import trendingai.shared.generated.resources.theme_follow_system
 import trendingai.shared.generated.resources.theme_light
@@ -149,7 +116,6 @@ import trendingai.shared.generated.resources.theme_light
 import trendingai.shared.generated.resources.favorites
 import trendingai.shared.generated.resources.feedback
 import trendingai.shared.generated.resources.feedback_desc
-import trendingai.shared.generated.resources.privacy_policy
 import trendingai.shared.generated.resources.subscribe_title
 import trendingai.shared.generated.resources.subscribe_desc
 import trendingai.shared.generated.resources.summary_language
@@ -167,6 +133,7 @@ import trendingai.shared.generated.resources.feedback_rate_limit
 import trendingai.shared.generated.resources.feedback_email_placeholder
 import trendingai.shared.generated.resources.feedback_email_invalid
 import trendingai.shared.generated.resources.version
+import trendingai.shared.generated.resources.version_tap_to_check
 import trendingai.shared.generated.resources.version_up_to_date
 
 @OptIn(ExperimentalMaterial3Api::class, ExperimentalMaterial3ExpressiveApi::class)
@@ -176,7 +143,8 @@ fun SettingsScreen(
     onNavigateToFavorites: () -> Unit = {},
     onNavigateToFeedback: () -> Unit = {},
     onNavigateToSubscribe: () -> Unit = {},
-    onNavigateToWebPage: (url: String, title: String) -> Unit = { _, _ -> }
+    onNavigateToAppearance: () -> Unit = {},
+    onNavigateToAbout: () -> Unit = {}
 ) {
     val isIos = isIosPlatform()
     val uriHandler = LocalUriHandler.current
@@ -191,9 +159,7 @@ fun SettingsScreen(
     val appVersion = remember { getAppVersion() }
     val isChecking by globalUpdateChecker.isChecking.collectAsState()
     val isUpToDate by globalUpdateChecker.isUpToDate.collectAsState()
-    var showDonateDialog by remember { mutableStateOf(false) }
     var showSummaryLanguageDialog by remember { mutableStateOf(false) }
-    var showOpenLinksDialog by remember { mutableStateOf(false) }
     // 语言采集流程：点「赞助 Pro」→ 采集期望语言（复用反馈接口提交）→ 成功后跳赞助页
     // 登录用户带上 GitHub 身份（便于与赞助对齐 + 后续通知）；未登录则收邮箱
     val authState by globalAuthManager.authState.collectAsState()
@@ -206,19 +172,6 @@ fun SettingsScreen(
     val settingsScope = rememberCoroutineScope()
     val permissionDeniedMsg = stringResource(Res.string.notification_permission_denied)
     val openSystemSettingsLabel = stringResource(Res.string.open_system_settings)
-
-    if (showOpenLinksDialog) {
-        AlertDialog(
-            onDismissRequest = { showOpenLinksDialog = false },
-            title = { Text(stringResource(Res.string.open_links_in_browser)) },
-            text = { Text(stringResource(Res.string.open_links_in_browser_message)) },
-            confirmButton = {
-                TextButton(onClick = { showOpenLinksDialog = false }) {
-                    Text(stringResource(Res.string.close))
-                }
-            }
-        )
-    }
 
     if (showSummaryLanguageDialog) {
         AlertDialog(
@@ -255,69 +208,6 @@ fun SettingsScreen(
         LanguageCaptureDialog(isLoggedIn = isLoggedIn, onDismiss = { showLangCaptureDialog = false })
     }
 
-    if (showDonateDialog) {
-        AlertDialog(
-            onDismissRequest = { showDonateDialog = false },
-            title = { Text(stringResource(Res.string.donate)) },
-            text = {
-                Column {
-                    Text(stringResource(Res.string.donate_message))
-                    Spacer(Modifier.height(16.dp))
-                    Row(
-                        verticalAlignment = Alignment.CenterVertically,
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .clickable {
-                                trackEvent("settings_donate_github")
-                                ProSponsor.openSponsorPage(ProSponsor.SOURCE_SETTINGS_DONATE)
-                            }
-                            .padding(vertical = 12.dp)
-                    ) {
-                        Icon(
-                            painter = githubLogoPainter(),
-                            contentDescription = null,
-                            tint = Color.Unspecified,
-                            modifier = Modifier.size(24.dp)
-                        )
-                        Column(modifier = Modifier.padding(start = 16.dp)) {
-                            Text("GitHub Sponsors", style = MaterialTheme.typography.bodyLarge)
-                            Text(
-                                text = stringResource(Res.string.donate_github_desc),
-                                style = MaterialTheme.typography.bodySmall,
-                                color = MaterialTheme.colorScheme.onSurfaceVariant
-                            )
-                        }
-                    }
-                    HorizontalDivider()
-                    Row(
-                        verticalAlignment = Alignment.CenterVertically,
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(vertical = 12.dp)
-                    ) {
-                        Icon(
-                            Icons.Default.VolunteerActivism,
-                            contentDescription = null,
-                            tint = MaterialTheme.colorScheme.onSurfaceVariant,
-                            modifier = Modifier.size(24.dp)
-                        )
-                        SelectionContainer(modifier = Modifier.padding(start = 16.dp)) {
-                            Text(
-                                text = stringResource(Res.string.donate_alipay, Constants.ALIPAY_ACCOUNT),
-                                style = MaterialTheme.typography.bodyLarge
-                            )
-                        }
-                    }
-                }
-            },
-            confirmButton = {
-                TextButton(onClick = { showDonateDialog = false }) {
-                    Text(stringResource(Res.string.confirm))
-                }
-            }
-        )
-    }
-
     Scaffold(
         topBar = {
             TopAppBar(
@@ -338,81 +228,25 @@ fun SettingsScreen(
         ) {
             // 分组 1: 个性化
             item { SettingsHeader(stringResource(Res.string.personalization)) }
+            // 外观：深色模式 + 主题色收进二级页，trailing 色点直观展示当前主题色
             item {
-                Column(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(horizontal = 16.dp, vertical = 12.dp)
-                ) {
-                    Row(
-                        verticalAlignment = Alignment.CenterVertically,
-                        modifier = Modifier.padding(bottom = 12.dp)
-                    ) {
-                        Icon(
-                            Icons.Default.Palette,
-                            null,
-                            tint = MaterialTheme.colorScheme.onSurfaceVariant
+                ListItem(
+                    headlineContent = { Text(stringResource(Res.string.appearance)) },
+                    supportingContent = { Text(themeModeText(themeMode)) },
+                    leadingContent = { Icon(Icons.Default.Palette, null) },
+                    trailingContent = {
+                        Box(
+                            modifier = Modifier
+                                .size(20.dp)
+                                .clip(CircleShape)
+                                .background(Color(seedColor))
                         )
-                        Text(
-                            text = stringResource(Res.string.dark_mode),
-                            style = MaterialTheme.typography.bodyLarge,
-                            modifier = Modifier.padding(start = 16.dp)
-                        )
+                    },
+                    modifier = Modifier.clickable {
+                        trackEvent("settings_appearance")
+                        onNavigateToAppearance()
                     }
-                    SingleChoiceSegmentedButtonRow(
-                        modifier = Modifier.fillMaxWidth()
-                    ) {
-                        ThemeMode.entries.forEachIndexed { index, mode ->
-                            val labelRes = when (mode) {
-                                ThemeMode.FOLLOW_SYSTEM -> Res.string.theme_follow_system
-                                ThemeMode.LIGHT -> Res.string.theme_light
-                                ThemeMode.DARK -> Res.string.theme_dark
-                            }
-                            SegmentedButton(
-                                selected = themeMode == mode,
-                                onClick = {
-                                    trackEvent("settings_theme_change", mapOf("theme" to mode.name.lowercase()))
-                                    globalSettingsManager.setThemeMode(mode)
-                                },
-                                shape = SegmentedButtonDefaults.itemShape(
-                                    index = index,
-                                    count = ThemeMode.entries.size
-                                ),
-                                label = {
-                                    Text(
-                                        text = stringResource(labelRes),
-                                        style = MaterialTheme.typography.labelMedium
-                                    )
-                                }
-                            )
-                        }
-                    }
-
-                    Spacer(Modifier.height(20.dp))
-
-                    Row(
-                        verticalAlignment = Alignment.CenterVertically,
-                        modifier = Modifier.padding(bottom = 12.dp)
-                    ) {
-                        Icon(
-                            Icons.Default.ColorLens,
-                            null,
-                            tint = MaterialTheme.colorScheme.onSurfaceVariant
-                        )
-                        Text(
-                            text = stringResource(Res.string.theme_color),
-                            style = MaterialTheme.typography.bodyLarge,
-                            modifier = Modifier.padding(start = 16.dp)
-                        )
-                    }
-                    SwatchGrid(
-                        selected = seedColor,
-                        onSelect = { seed ->
-                            trackEvent("settings_seed_color", mapOf("seed" to seed.id))
-                            globalSettingsManager.setSeedColor(seed.argb)
-                        }
-                    )
-                }
+                )
             }
             // 我的收藏
             item {
@@ -441,25 +275,21 @@ fun SettingsScreen(
 
             // 分组 2: 应用设置
             item { SettingsHeader(stringResource(Res.string.app_settings)) }
+            // 语言设置：trailing 切应用语言；行点击弹摘要语言说明（原独立条目并入此处）
             item {
-                if (isIos) {
-                    ListItem(
-                        headlineContent = { Text(stringResource(Res.string.language_settings)) },
-                        supportingContent = { Text(stringResource(Res.string.language_system_follow)) },
-                        trailingContent = {
+                ListItem(
+                    headlineContent = { Text(stringResource(Res.string.language_settings)) },
+                    supportingContent = { Text(stringResource(Res.string.summary_language_desc)) },
+                    leadingContent = { Icon(Icons.Default.Language, null) },
+                    trailingContent = {
+                        if (isIos) {
                             Text(
                                 text = stringResource(Res.string.open_system_settings),
-                                color = MaterialTheme.colorScheme.primary
+                                color = MaterialTheme.colorScheme.primary,
+                                modifier = Modifier.clickable { openAppSettings() }
                             )
-                        },
-                        leadingContent = { Icon(Icons.Default.Language, null) },
-                        modifier = Modifier.clickable { openAppSettings() }
-                    )
-                } else {
-                    var expanded by remember { mutableStateOf(false) }
-                    ListItem(
-                        headlineContent = { Text(stringResource(Res.string.language_settings)) },
-                        trailingContent = {
+                        } else {
+                            var expanded by remember { mutableStateOf(false) }
                             Box {
                                 Text(
                                     text = languageOptionText(appLanguage),
@@ -482,58 +312,11 @@ fun SettingsScreen(
                                     }
                                 }
                             }
-                        },
-                        leadingContent = { Icon(Icons.Default.Language, null) }
-                    )
-                }
-            }
-            // 摘要语言说明
-            item {
-                ListItem(
-                    headlineContent = { Text(stringResource(Res.string.summary_language)) },
-                    supportingContent = { Text(stringResource(Res.string.summary_language_desc)) },
-                    leadingContent = { Icon(Icons.Default.Translate, null) },
+                        }
+                    },
                     modifier = Modifier.clickable {
                         trackEvent("settings_summary_language", mapOf("app_language" to appLanguage.name.lowercase()))
                         showSummaryLanguageDialog = true
-                    }
-                )
-            }
-            // 外链打开方式
-            item {
-                ListItem(
-                    headlineContent = { Text(stringResource(Res.string.open_links_in_browser)) },
-                    supportingContent = { Text(stringResource(Res.string.open_links_in_browser_desc)) },
-                    leadingContent = { Icon(Icons.Default.OpenInBrowser, null) },
-                    trailingContent = {
-                        Switch(
-                            checked = openLinksInCustomTab,
-                            onCheckedChange = { enabled ->
-                                trackEvent("settings_open_links_in_browser", mapOf("enabled" to enabled.toString()))
-                                globalSettingsManager.setOpenLinksInCustomTab(enabled)
-                            }
-                        )
-                    },
-                    modifier = Modifier.clickable {
-                        trackEvent("settings_open_links_detail")
-                        showOpenLinksDialog = true
-                    }
-                )
-            }
-            // 「只看 New」默认开关：只决定进入 app 时的初始状态，榜单页手动切换不回写
-            item {
-                ListItem(
-                    headlineContent = { Text(stringResource(Res.string.new_only_default)) },
-                    supportingContent = { Text(stringResource(Res.string.new_only_default_desc)) },
-                    leadingContent = { Icon(Icons.Default.FiberNew, null) },
-                    trailingContent = {
-                        Switch(
-                            checked = trendingNewOnlyDefault,
-                            onCheckedChange = { enabled ->
-                                trackEvent("settings_new_only_default", mapOf("enabled" to enabled.toString()))
-                                globalSettingsManager.setTrendingNewOnlyDefault(enabled)
-                            }
-                        )
                     }
                 )
             }
@@ -568,6 +351,42 @@ fun SettingsScreen(
                             }
                         }
                     }
+                )
+            }
+            // 「只看 New」默认开关：只决定进入 app 时的初始状态，榜单页手动切换不回写
+            item {
+                ListItem(
+                    headlineContent = { Text(stringResource(Res.string.new_only_default)) },
+                    supportingContent = { Text(stringResource(Res.string.new_only_default_desc)) },
+                    leadingContent = { Icon(Icons.Default.FiberNew, null) },
+                    trailingContent = {
+                        Switch(
+                            checked = trendingNewOnlyDefault,
+                            onCheckedChange = { enabled ->
+                                trackEvent("settings_new_only_default", mapOf("enabled" to enabled.toString()))
+                                globalSettingsManager.setTrendingNewOnlyDefault(enabled)
+                            }
+                        )
+                    }
+                )
+            }
+            // 外链打开方式：行点击与开关同为切换，说明并入 supporting，不再弹对话框
+            item {
+                val toggleOpenLinks = { enabled: Boolean ->
+                    trackEvent("settings_open_links_in_browser", mapOf("enabled" to enabled.toString()))
+                    globalSettingsManager.setOpenLinksInCustomTab(enabled)
+                }
+                ListItem(
+                    headlineContent = { Text(stringResource(Res.string.open_links_in_browser)) },
+                    supportingContent = { Text(stringResource(Res.string.open_links_in_browser_desc)) },
+                    leadingContent = { Icon(Icons.Default.OpenInBrowser, null) },
+                    trailingContent = {
+                        Switch(
+                            checked = openLinksInCustomTab,
+                            onCheckedChange = toggleOpenLinks
+                        )
+                    },
+                    modifier = Modifier.clickable { toggleOpenLinks(!openLinksInCustomTab) }
                 )
             }
             // 每日精选提醒：本地定时通知（WorkManager），全渠道可用；iOS 未支持则隐藏
@@ -625,24 +444,41 @@ fun SettingsScreen(
                     }
                 )
             }
-            // apk 渠道显示自建更新检查，iOS 跳转官网；play 渠道由商店管理更新，不显示
-            if (globalUpdateChecker.isEnabled || isIos) {
-                item {
-                    ListItem(
-                        headlineContent = { Text(stringResource(Res.string.check_updates)) },
-                        trailingContent = {
-                            when {
-                                !isIos && isChecking -> LoadingIndicator(
-                                    modifier = Modifier.size(24.dp)
-                                )
-                                !isIos && isUpToDate -> Text(
-                                    stringResource(Res.string.version_up_to_date),
-                                    color = MaterialTheme.colorScheme.primary
-                                )
-                            }
-                        },
-                        leadingContent = { Icon(Icons.Default.Refresh, null) },
-                        modifier = Modifier.clickable(enabled = !isChecking) {
+            // 关于我们：官网/联系作者/捐赠/隐私政策收进二级页
+            item {
+                ListItem(
+                    headlineContent = { Text(stringResource(Res.string.about_us)) },
+                    supportingContent = { Text(stringResource(Res.string.about_us_desc)) },
+                    leadingContent = { Icon(Icons.Default.Info, null) },
+                    modifier = Modifier.clickable {
+                        trackEvent("settings_about")
+                        onNavigateToAbout()
+                    }
+                )
+            }
+            // 版本 + 检查更新合并为一行：apk 渠道点按自建检查，iOS 跳官网，play 渠道仅展示版本号
+            item {
+                val canCheckUpdate = globalUpdateChecker.isEnabled || isIos
+                ListItem(
+                    headlineContent = { Text(stringResource(Res.string.version)) },
+                    supportingContent = if (canCheckUpdate) {
+                        { Text(stringResource(Res.string.version_tap_to_check)) }
+                    } else null,
+                    leadingContent = { Icon(Icons.Default.Numbers, null) },
+                    trailingContent = {
+                        when {
+                            !isIos && isChecking -> LoadingIndicator(
+                                modifier = Modifier.size(24.dp)
+                            )
+                            !isIos && isUpToDate -> Text(
+                                stringResource(Res.string.version_up_to_date),
+                                color = MaterialTheme.colorScheme.primary
+                            )
+                            else -> Text(appVersion, color = MaterialTheme.colorScheme.outline)
+                        }
+                    },
+                    modifier = if (canCheckUpdate) {
+                        Modifier.clickable(enabled = !isChecking) {
                             trackEvent("settings_check_update")
                             if (isIos) {
                                 uriHandler.openUri(Constants.OFFICIAL_WEBSITE_URL)
@@ -650,57 +486,7 @@ fun SettingsScreen(
                                 globalUpdateChecker.manualCheck()
                             }
                         }
-                    )
-                }
-            }
-            item {
-                ListItem(
-                    headlineContent = { Text(stringResource(Res.string.about_us)) },
-                    supportingContent = { Text(stringResource(Res.string.about_us_desc)) },
-                    leadingContent = { Icon(Icons.Default.Info, null) },
-                    modifier = Modifier.clickable {
-                        uriHandler.openUri(Constants.OFFICIAL_WEBSITE_URL)
-                    }
-                )
-            }
-            item {
-                ListItem(
-                    headlineContent = { Text(stringResource(Res.string.contact_author)) },
-                    trailingContent = {
-                        SelectionContainer {
-                            Text(Constants.AUTHOR_EMAIL, color = MaterialTheme.colorScheme.outline)
-                        }
-                    },
-                    leadingContent = { Icon(Icons.Default.AlternateEmail, null) }
-                )
-            }
-            item {
-                ListItem(
-                    headlineContent = { Text(stringResource(Res.string.donate)) },
-                    leadingContent = { Icon(Icons.Default.VolunteerActivism, null) },
-                    modifier = Modifier.clickable {
-                        trackEvent("settings_donate")
-                        showDonateDialog = true
-                    }
-                )
-            }
-            item {
-                val privacyTitle = stringResource(Res.string.privacy_policy)
-                ListItem(
-                    headlineContent = { Text(privacyTitle) },
-                    leadingContent = { Icon(Icons.Default.PrivacyTip, null) },
-                    modifier = Modifier.clickable {
-                        onNavigateToWebPage(Constants.PRIVACY_POLICY_URL, privacyTitle)
-                    }
-                )
-            }
-            item {
-                ListItem(
-                    headlineContent = { Text(stringResource(Res.string.version)) },
-                    trailingContent = {
-                        Text(appVersion, color = MaterialTheme.colorScheme.outline)
-                    },
-                    leadingContent = { Icon(Icons.Default.Numbers, null) }
+                    } else Modifier
                 )
             }
         }
@@ -718,67 +504,21 @@ private fun languageOptionText(language: AppLanguage): String {
 }
 
 @Composable
+private fun themeModeText(mode: ThemeMode): String {
+    val labelRes = when (mode) {
+        ThemeMode.FOLLOW_SYSTEM -> Res.string.theme_follow_system
+        ThemeMode.LIGHT -> Res.string.theme_light
+        ThemeMode.DARK -> Res.string.theme_dark
+    }
+    return stringResource(labelRes)
+}
+
+@Composable
 private fun homeTabOptionText(tab: HomeTab): String = when (tab) {
     HomeTab.GitHub -> "GitHub"
     HomeTab.HackerNews -> stringResource(Res.string.hackernews_title)
     HomeTab.ProductHunt -> stringResource(Res.string.producthunt_title)
     HomeTab.Picks -> stringResource(Res.string.picks_title)
-}
-
-@OptIn(ExperimentalLayoutApi::class)
-@Composable
-private fun SwatchGrid(
-    selected: Long,
-    onSelect: (ThemeSeed) -> Unit,
-) {
-    FlowRow(
-        modifier = Modifier
-            .fillMaxWidth()
-            .selectableGroup(),
-        horizontalArrangement = Arrangement.spacedBy(12.dp),
-        verticalArrangement = Arrangement.spacedBy(12.dp),
-    ) {
-        PRESET_PALETTE.forEach { seed ->
-            ThemeSwatch(
-                seed = seed,
-                selected = seed.argb == selected,
-                onClick = { onSelect(seed) },
-            )
-        }
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun ThemeSwatch(
-    seed: ThemeSeed,
-    selected: Boolean,
-    onClick: () -> Unit,
-) {
-    val color = Color(seed.argb)
-    val name = stringResource(seed.nameRes)
-    Surface(
-        selected = selected,
-        onClick = onClick,
-        shape = CircleShape,
-        color = color,
-        modifier = Modifier
-            .size(40.dp)
-            .semantics {
-                contentDescription = name
-                role = Role.RadioButton
-            },
-    ) {
-        if (selected) {
-            Box(contentAlignment = Alignment.Center, modifier = Modifier.fillMaxSize()) {
-                Icon(
-                    imageVector = Icons.Default.Check,
-                    contentDescription = null,
-                    tint = if (color.luminance() < 0.5f) Color.White else Color.Black,
-                )
-            }
-        }
-    }
 }
 
 @Composable
